requirements.txt
# Base
Django>=6.0,<7.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Filtering
django-filter>=24.0,<25.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Security Headers
django-csp>=3.8,<4.0
django-permissions-policy>=4.18,<5.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0

# Production server
gunicorn>=21.0,<23.0

# Environment variables
python-decouple>=3.8,<4.0
